Finished Programming First Draft Mission Tree Norway
I have finished the first draft of the Norwegian mission tree, though no localisation yet and still missing a few of the newer ideas. I need to add some unqiue gov forms, monuments, event chains and the religious missions, aswell as change the borders of a few of the american missions.